refactor input tabs
